Fix errors in unit tests due to refactoring
