--- tests/org.eclipse.cdt.lsp.clangd.tests/src/org/eclipse/cdt/lsp/clangd/editor/tests/EditorTestUtils.java
@@ -1,51 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2023 Bachmann electronic GmbH and others.
- *
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     Gesa Hentschke (Bachmann electronic GmbH) - initial implementation
- *     Alexander Fedorov (ArSysOp) - extract headless part
- *******************************************************************************/
-
-package org.eclipse.cdt.lsp.clangd.editor.tests;
-
-import java.net.URI;
-
-import org.eclipse.core.resources.IFile;
-import org.eclipse.ui.IEditorPart;
-import org.eclipse.ui.IWorkbenchPage;
-import org.eclipse.ui.IWorkbenchWindow;
-import org.eclipse.ui.PartInitException;
-import org.eclipse.ui.PlatformUI;
-import org.eclipse.ui.ide.IDE;
-
-public final class EditorTestUtils {
-
-	public static IEditorPart openInEditor(URI uri, String editorID) throws PartInitException {
-		IEditorPart part = IDE.openEditor(EditorTestUtils.getWorkbenchPage(), uri, editorID, true);
-		part.setFocus();
-		return part;
-	}
-
-	public static IEditorPart openInEditor(IFile file) throws PartInitException {
-		IEditorPart part = IDE.openEditor(EditorTestUtils.getWorkbenchPage(), file);
-		part.setFocus();
-		return part;
-	}
-
-	public static boolean closeEditor(IEditorPart editor, boolean save) {
-		IWorkbenchWindow workbenchWindow = PlatformUI.getWorkbench().getActiveWorkbenchWindow();
-		IWorkbenchPage page = workbenchWindow.getActivePage();
-		return page.closeEditor(editor, save);
-	}
-
-	private static IWorkbenchPage getWorkbenchPage() {
-		return PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage();
-	}
-
-}
