+++ tests/org.eclipse.cdt.lsp.server.test/fragment.xml
@@ -4,7 +4,7 @@
    <extension
          point="org.eclipse.cdt.lsp.serverProvider">
       <server
-            class="org.eclipse.cdt.lsp.test.server.MockCLanguageServerProvider"
+            class="org.eclipse.cdt.lsp.server.test.MockCLanguageServerProvider"
             priority="normal">
       </server>
    </extension>
